refactored the code and added component scan pathing for service and repo folders
