+++ src/main/java/com/loot/server/ServerApplication.java
@@ -4,7 +4,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
-@ComponentScan(basePackages = {"com.loot.server.api", "com.loot.server.socket"})
+@ComponentScan(basePackages = {"com.loot.server.api", "com.loot.server.socket", "com.loot.server.service", "com.loot.server.repositories"})
 @SpringBootApplication
 public class ServerApplication {
