+++ src/main/java/com/loot/server/domain/Player.java
@@ -1,4 +1,4 @@
-package com.loot.server.socket.domain;
+package com.loot.server.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
