+++ src/main/java/com/loot/server/service/GameService.java
@@ -1,4 +1,4 @@
-package com.loot.server.api.service;
+package com.loot.server.service;
 
 import com.loot.server.domain.GameCreationDto;
