+++ src/main/java/com/loot/server/service/impl/GameServiceImpl.java
@@ -1,11 +1,11 @@
-package com.loot.server.api.service.impl;
+package com.loot.server.service.impl;
 
 import java.util.HashSet;
 import java.util.Random;
 import java.util.Set;
 
 import com.loot.server.domain.GameCreationDto;
-import com.loot.server.api.service.GameService;
+import com.loot.server.service.GameService;
 import org.springframework.stereotype.Service;
 
 @Service
